fix(#899): apply fix pass changes
